chore(backend): restrict token introspection middleware access actions
