--- packages/backend/src/open_payments/auth/grant.ts
@@ -1,15 +1,0 @@
-// TODO: replace with open-payments generated types
-export enum AccessType {
-  IncomingPayment = 'incoming-payment',
-  OutgoingPayment = 'outgoing-payment',
-  Quote = 'quote'
-}
-
-export enum AccessAction {
-  Create = 'create',
-  Read = 'read',
-  ReadAll = 'read-all',
-  Complete = 'complete',
-  List = 'list',
-  ListAll = 'list-all'
-}
