+++ packages/backend/src/open_payments/auth/service.test.ts
@@ -3,8 +3,8 @@ import nock, { Definition } from 'nock'
 import { URL } from 'url'
 import { v4 as uuid } from 'uuid'
 
-import { AccessType, AccessAction } from './grant'
 import { AuthService, TokenInfo } from './service'
+import { AccessType, AccessAction } from '../grant/model'
 import { Config } from '../../config/app'
 import { IocContract } from '@adonisjs/fold'
 import { initIocContainer } from '../../'
