+++ packages/backend/src/open_payments/auth/service.ts
@@ -1,7 +1,7 @@
 import axios from 'axios'
 import { Logger } from 'pino'
 
-import { AccessAction, AccessType } from './grant'
+import { AccessAction, AccessType } from '../grant/model'
 import { OpenAPI, HttpMethod, ResponseValidator } from 'openapi'
 import { JWK } from 'http-signature-utils'
 import { AmountJSON } from '../amount'
