+++ packages/backend/src/open_payments/grant/model.ts
@@ -1,8 +1,23 @@
 import { Model, QueryContext } from 'objection'
 
-import { AccessType, AccessAction } from '../auth/grant'
 import { AuthServer } from '../authServer/model'
 import { BaseModel } from '../../shared/baseModel'
+
+// TODO: replace with open-payments generated types
+export enum AccessType {
+  IncomingPayment = 'incoming-payment',
+  OutgoingPayment = 'outgoing-payment',
+  Quote = 'quote'
+}
+
+export enum AccessAction {
+  Create = 'create',
+  Read = 'read',
+  ReadAll = 'read-all',
+  Complete = 'complete',
+  List = 'list',
+  ListAll = 'list-all'
+}
 
 export class Grant extends BaseModel {
   public static get tableName(): string {
