+++ packages/backend/src/open_payments/grant/service.test.ts
@@ -2,9 +2,8 @@ import { IocContract } from '@adonisjs/fold'
 import { faker } from '@faker-js/faker'
 import { Knex } from 'knex'
 
-import { Grant } from './model'
+import { AccessType, AccessAction, Grant } from './model'
 import { GrantOptions, GrantService } from './service'
-import { AccessType, AccessAction } from '../auth/grant'
 import { AuthServer } from '../authServer/model'
 import { initIocContainer } from '../..'
 import { AppServices } from '../../app'
